+++ frontend/src/main.tsx
@@ -4,13 +4,16 @@ import { Provider } from "react-redux";
 import App from "./App";
 import { store } from "./app/store";
 import { ThemeProvider } from "./theme/ThemeProvider";
+import { BrowserRouter } from "react-router-dom";
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
     <Provider store={store}>
       <ThemeProvider>
-        <App />
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
       </ThemeProvider>
     </Provider>
   </React.StrictMode>